feat: add api create book

+++ routes/api.php
@@ -6,6 +6,7 @@ $router->get('/', App\Controllers\Home\GetHomePage::class);
 
 $router->group('/api/v1/', function ($router) {
     $auth = App\Middlewares\Jwt::class;
+    $admin = App\Middlewares\Admin::class;
 
     $router->get('/', App\Controllers\Home\GetHomePage::class);
 
@@ -24,12 +25,16 @@ $router->group('/api/v1/', function ($router) {
         $router->post('/', App\Controllers\Contact\CreateContactController::class);
     });
 
-    $router->group('/category', function ($router) use ($auth) {
+    $router->group('/category', function ($router) use ($auth, $admin) {
         $router->get('/dropdown', App\Controllers\Category\AllDropdownCategoryController::class, [$auth]);
-        $router->get('/', App\Controllers\Category\AllCategoryController::class, [$auth]);
-        $router->post('/', App\Controllers\Category\CreateCategoryController::class, [$auth]);
-        $router->get('/{id}', App\Controllers\Category\DetailCategoryController::class, [$auth]);
-        $router->put('/{id}', App\Controllers\Category\UpdateCategoryController::class, [$auth]);
-        $router->delete('/{id}', App\Controllers\Category\DeleteCategoryController::class, [$auth]);
-     });
+        $router->get('/', App\Controllers\Category\AllCategoryController::class, [$auth, $admin]);
+        $router->post('/', App\Controllers\Category\CreateCategoryController::class, [$auth, $admin]);
+        $router->get('/{id}', App\Controllers\Category\DetailCategoryController::class, [$auth, $admin]);
+        $router->put('/{id}', App\Controllers\Category\UpdateCategoryController::class, [$auth, $admin]);
+        $router->delete('/{id}', App\Controllers\Category\DeleteCategoryController::class, [$auth, $admin]);
+    });
+
+    $router->group('/book', function ($router) use ($auth, $admin) {
+        $router->post('/', App\Controllers\Book\CreateBookController::class, [$auth, $admin]);
+    });
 });
